remove: remove `constants/float16/lanczos-gamma-g`
The value added in #9022 is not appropriate for half-precision
numerical approximation. As observed for the equivalent `float64`
and `float32` constants, the value we use for `g` depends on what
approximation we use in our gamma function implementations. In
general, we should only add this constant if and when we add a
half-precision gamma approximation. Something we are not likely
to do anytime soon.

BREAKING CHANGE: remove `constants/float16/lanczos-gamma-g`

This feature was only recently added and has not been published.
As such, the impact of removing this constant should be negligible
for downstream users.

Ref: https://github.com/stdlib-js/stdlib/pull/9022

---
type: pre_commit_static_analysis_report
description: Results of running static analysis checks when committing changes.
report:
  - task: lint_filenames
    status: passed
  - task: lint_editorconfig
    status: passed
  - task: lint_markdown
    status: na
  - task: lint_package_json
    status: na
  - task: lint_repl_help
    status: na
  - task: lint_javascript_src
    status: na
  - task: lint_javascript_cli
    status: na
  - task: lint_javascript_examples
    status: na
  - task: lint_javascript_tests
    status: na
  - task: lint_javascript_benchmarks
    status: na
  - task: lint_python
    status: na
  - task: lint_r
    status: na
  - task: lint_c_src
    status: na
  - task: lint_c_examples
    status: na
  - task: lint_c_benchmarks
    status: na
  - task: lint_c_tests_fixtures
    status: na
  - task: lint_shell
    status: na
  - task: lint_typescript_declarations
    status: passed
  - task: lint_typescript_tests
    status: na
  - task: lint_license_headers
    status: passed
---